fix return type of extectedSize method
size_t is 32 bit unsigned number on 32 bit architectures,
this is not enough for when download job has more than 4 GiB

+++ libosmscout-client-qt/include/osmscout/MapManager.h
@@ -67,7 +67,7 @@ public:
     return map.getPath();
   }
 
-  inline size_t expectedSize() const override
+  inline uint64_t expectedSize() const override
   {
     return map.getSize();
   }
